requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
Flask==2.3.3
requests==2.31.0
nltk==3.8.1
spacy==3.7.2
pymongo==4.6.1

